Fix image_url selection and add reboot on success
the command line image_url selection forced the installer to try
download an image prior to the network being up, which led to failures.
Gate the service starting on the initqueue service finishing to
guarantee that it all starts up properly.

Also, on success, inform the user and reboot the system

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>