Wait for device, drop dependency on initrd-root-device
The dependency on `initrd-root-device.target` in `coreos-installer.target`
can't do anything useful - for the installer, there
*isn't* a root device from the perspective of systemd/dracut bootup.
Instead, its goal is to just run custom code to write out to the
targeted block device.

Having this dependency here means that any units queued by the
"regular" initramfs that end up as part of `initrd-root-device.target`
will be executed during the install phase, which is definitely not
desired.