systemd: set SYSTEMD_SULOGIN_FORCE=1 for emergency.service
On error of the installer we invoke emergency.target. The executable
that gets called in emergency.service will won't give us a shell if
the root account is locked. We'll get a message like this:

```
Cannot open access to console, the root account is locked.
See sulogin(8) man page for more details.
```

We can force to login by setting SYSTEMD_SULOGIN_FORCE=1.