Just retry image fetches 5 times
OpenShift CI hit a provisioning failure in Packet.

Reading `man curl`, it's not just code `22` we need to retry on
but also things like:

- `7      Failed to connect to host.`
- `18     Partial file. Only a part of the file was transferred.`
- `92     Stream error in HTTP/2 framing layer.`

And the man page explicitly says others could be added in the future.
We could make this nicer but this is the legacy branch.

Let's just retry unconditionally a few times.