Stop discarding disk contents
In 7b89735c0906 we started issuing block discard requests over the
destination disk, before install and after install failure, as a
courtesy to SSD wear levelers and LVM thin pools.  However, this
assumes that the entire disk is ours to discard.  Users might want to
reprovision an existing machine without disturbing a data partition
they've created on the boot disk; their Ignition config can then
recreate the data partition in the same place and reuse the filesystem
inside.

Discard the discard functionality for now.  We shouldn't require users
to explicitly specify --dont-delete-my-data, and the semantics of
--delete-my-data (under whatever name, e.g. --discard-device) would
be difficult to explain relative to the value they provide.  Users can
still run blkdiscard(8) before coreos-installer if they need that
functionality.

Fixes #170.