source: keep going on signature-fetching errors
This tweaks the signature fetching logic, so that it log-and-continues
on any kind of errors (e.g. timeouts or broken redirects).
This is mostly useful for the "insecure" mode, where remote images
can be consumed even without a corresponding signature.
In "normal" mode, such case will instead result in an hard-failure
when trying to validate the (missing) signature.
Additionally, this prints more error details on non-OK HTTP responses.

Ref: https://github.com/coreos/fedora-coreos-tracker/issues/402