download: Avoid using `is_some()`
I think using that API call is a bit of a code smell; it isn't
*wrong* but I think it's better to use matching like `if let Some(x)`.
Among other things it avoids an `unwrap()` call which is *definitely*
a code smell.  It's harder for the compiler to optimize out the
`unwrap()` too since it has to reason about the `is_some()` logic.

If agreed I'll change a few other places like this too.