ci: add offline testing
This turns on CI testing for the offline path (via osmet). In the
future, we will drop `--osmet` entirely since it'll always be on. (And
similarly, the `iso-offline-install` might become the same thing as
`iso-install` so we might be able to drop it entirely.)

See https://github.com/coreos/coreos-assembler/pull/1244.