pack: Support --real-rootdev for RHCOS+LUKS
Pairs with: https://github.com/coreos/coreos-assembler/pull/1483

We want RHCOS to support offline installs too (it's key to the UX)
but the LUKS container complicates things.  We discussed changing
the RHCOS LUKS bits to save the header in e.g. `/boot` but it's
a big change to that code, and introduces yet another layout
when we really want to switch to Ignition+LUKS.