systemd: coreos-installer-service: Add support for persisting coreos.autologin additional arg
Added support for an autologin cmdline arg which would be really useful to debug cases where
the ability to login to the machine through ssh is broken due to some networking issue. This is
useful for debugging certain cases like OCP UPI deploys as we don't have to tinker with
the ignition config. The command line arg can be added and the system rebooted to autologin
and diagnose the issue.