osmet: replace deprecated bincode::Config with bincode::Options
bincode::Options doesn't implement Copy and its methods move the Options
into the callee.  Add a helper function to create the correct Options
and just call it everywhere.

This changes from fixint to varint encoding, which is a compatibility
break in exchange for perhaps smaller file size.  Since we don't need
to be compatible with old osmet files, this should be okay.