Move logic to `libcoreinst` library
Create a new `libcoreinst` library which contains all the logic, and
make `coreos-installer` simply consist of `main.rs` which calls into it.

This is prep for adding another binary which will reuse code in
`libcoreinst`. Doing it this way is a bit awkward, because e.g.
`cmdline` is also in the library, but it allows minimal churn while
still making it easy to add other binaries.