rdcore: new binary for initrd CoreOS tasks
By design, FCOS and RHCOS have a lot going on in the initramfs stage and
the logic there is likely going to keep growing. As a result, we have
many non-trivial and crucial shell scripts living in the source config
(https://github.com/coreos/fedora-coreos-config) which (1) have poor
error-handling, (2) aren't easy to manually test or unit test, and (3)
have low visibility because they're deeply nested within some dracut
module.

I'd like to propose a new `rdcore` binary which will try to address
these issues simply by virtue of being in a proper programming language.
This is shamelessly intended to be a "kitchen sink" collection of
whatever logic is needed in the initramfs.

Down the line we can split this out into its own package. But for now,
it's much easier to tag along an existing one for experimenting without
having to incur too much paperwork. It's an awkward arrangement, but I
think the pros outweigh the cons overall, and at least
`coreos-installer` already isn't very distro-agnostic.

As for why make this a separate binary instead of just extending the
CLI, a few reasons:
1. it'll be easier to eventually rip out and move to its own package
2. thanks to LTO, we only pay for the bits of the library we use
3. we don't inadvertently add back installer support into the initrd