rdcore: add stream-hash subcommand
This command copies a stream from stdin to stdout, verifying the data via
SHA-256 as it goes.  It receives an input file in the following format:

stream-hash sha256 <chunk-size>
<hexdigest>
<hexdigest>
<hexdigest>
...

The idea is that we can read <chunk-size> bytes into RAM, check their
digest, send them to the next program in a shell pipeline, and move on
to the next hash.  This lets us safely unpack a streaming download while
verifying it, and is intended for downloading the PXE rootfs image.