download: use buffered I/O with 256 KiB buffer
The current image pipeline is:

    local file or HTTP response ->
    [some Readers that don't affect I/O size] ->
    BufReader with 8 KiB buffer ->
    decompressor ->
    copy() with 256 KiB buffer ->
    disk

copy()'s large buffer doesn't matter here.  The decompressor will read
data in whatever chunk size it wants, which will be aggregated only up
to 8 KiB.  The decompressor will also produce data in whatever chunk size
it wants, depending on the compressed stream, and those chunks won't be
aggregated at all.

Upgrade the BufReader to a 256 KiB buffer and add 256 KiB BufWriters.
This doesn't seem to affect performance much in my local tests, but it's
still the right thing to do.