io: drop copy() reimplementation
The right approach to increasing buffer sizes is to put a large
BufReader/BufWriter on the input/output.  The buffer size of the copy()
itself doesn't matter in a long pipeline, since the other components
may produce/consume data in smaller chunks.  At best it might hide the
lack of buffering at the ends.

Also add some buffers on large copies.