Add `util` module with command helpers
The pattern of creating a command and decoding its output is common
enough that it's about time we should have a helper for it.

Add a new `cmd_output()` which takes a command, runs it, and extracts
its output. Add `runcmd!()` and `runcmd_output!()` helpers to make it
easier to create and run a command.