install: make stash file the same size as the install disk
gptman currently allows partitions larger than the target disk, but
that's arguably a bug.  Don't depend on that behavior in a last-resort
error path.  We could calculate the minimum file size to contain the
saved partitions, but unallocated disk blocks are cheap, so let's just
size the stash file to match the original disk.