blockdev: have tests size GPTs just large enough for partitions
At runtime, we save partitions from a large disk into a GPT configured
for the smaller install image.  In the tests, we were creating both base
and image disks with the same size.  Change make_disk() to size the
disk just large enough for its partitions, create the GPT, and resize
to 10 GB afterward.