iso: use copy_file_range to copy the ISO if possible
On filesystems and kernels that support it, use `copy_file_range` to
copy the ISO file. CoW semantics is a good fit here because the
transformations we do on the output file are minimal anyway, so there
isn't really a question of balancing against delayed copy costs.

This brings down the time to inject an Ignition config from 4s to 50ms
on my computer.

To do this, use `openat-ext`'s `FileExt` trait which supports doing
`copy_file_range` on `File` objects. (`std::fs::copy()` also knows
`copy_file_range`, but it only takes file paths, so we'd lose the
`O_EXCL | O_CREAT` semantics we want here).