blockdev: verify that partitions can be restored before installing
If the combined partition table is inconsistent, causing partition table
restore to fail, we'll clear the partition table and restore just the
saved partitions.  But if the original partition table is inconsistent,
e.g. two saved partitions with the same GUID, the second restore will
also fail and we'll drop the saved partitions on the floor.  To avoid
this, dry-run GPT.write_into() into a temporary file immediately after
saving, allowing us to bail out before the install starts.