blockdev: call GPTHeader.update_from() on install image GPT
gptman 0.7 no longer updates the header from disk in GPT.write_from().
SavedPartitions.merge() is the only place where we write a GPT we didn't
create ourselves; call GPTHeader.update_from() there to update the
header's disk size from the size of the install image to the size of the
full disk.