iso: add `kargs` commands to manage kernel args
The new commands are `coreos-installer iso kargs modify/reset/show`.
These allow modifying the baked in kernel arguments in a CoreOS live
ISO.

My primary motivation for this is for embedding the right serial console
argument into the ISO before booting it:

```
$ coreos-installer iso kargs modify --append console=ttyS0 my.iso
$ coreos-installer iso kargs show my.iso
mitigations=auto,nosmt ... console=ttyS0
$ cosa run --qemu-iso my.iso -c
...
[    0.000000] Command line: BOOT_IMAGE=/images/vmlinuz ...  console=ttyS0
...
```

This example specifically demonstrates what cosa will do automatically
itself once it can rely on this functionality. This will for example
allow `testiso` tests to be much more useful because we'll actually have
console logs.

But it's of course useful generically. Another way to look at it is that
it mirrors/brings to parity the provisioning capabilities of the live
PXE path.

Co-authored-by: Nikita Dubrovskii <nikita@linux.ibm.com>