rdcore/rootmap: also inject rootflags karg
Because we now always inject a `root` karg, it means that it's systemd
which does the mount. But then, systemd doesn't know to use the
`prjquota` flag we want by default for xfs.

Call out to a `coreos-rootflags` helper where the source of truth for
default root mount flags lives and embed its output if any is provided.