install: add helper to visit just the BLS `options`
Really, all we care about right now in the BLS config is the `options`
key, which is where the kargs are saved. Let's add a function for this
use case which just wraps `visit_bls_entry`. Switch the two callers to
use that.

This for example makes `bls_entry_write_platform` more precise because
now it's only doing its blind `sed` on the kargs instead of on the whole
config.