install: gracefully no-op when handling kargs
The callers of this function right now do always pass non-None args when
calling it, so it doesn't matter much. But it'll make new code in the
next patch cleaner by being able to just call this function
unconditionally, even if there's no kargs to modify (i.e. the user is
just querying).