blockdev: use --nodeps when querying single device
`lsblk` by default outputs the target device followed by all its reverse
dependencies. But interestingly, when using `--pairs`, the `lsblk` in
RHCOS seems to output the reverse dependency first before the target
device.

We were relying on correct ordering in `lsblk_single`, where all we care
about is the target device. Instead, explicitly use `--nodeps` in that
case so that `lsblk` knows to only print the target device. (The
`lsblk()` helper actually used to support that, but it was removed in
88de4fcc.)

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1916502