iso: refactor ContentFile as an enum; extend for stdout streaming
Explicitly separate the in-place and copied behaviors.  Use NamedTempFile
for the copied case, rather than writing directly to the final location
and deleting the output file on failure.  Add a case to support streaming
directly to stdout.

This will set the copied output file mode to 0600, which seems okay
because the embedded Ignition config may contain secrets.  It also ensures
that the permanent filename is not used until the file is completely
written.