blockdev: Resolve symlink disk paths when creating Disk objects
Sometimes disks may be given by a path that looks something
like `/dev/disk/by-id/wwn-0x5555555555555555`. This path is
a symlink to the actual `/dev/*` device.

This commit prevents a confusing error message from showing
when the installer is given a busy disk that is a symlink to
a `/dev/*` device and not an actual `/dev/*` directly.

The solution used here is simply resolving the symlink before
creating a `Disk` object. This is done because code later down
the line assumes that the path is going to have a `/dev/*` format.

Closes: #512