iso: add minimal ISO9660 parser
We do a lot of sophisticated things with our live ISOs. An upcoming
patch will push things yet further by supporting embedding in the ISO a
compressed minimal version of itself.

Currently, these modifications (together with Ignition embedding and
kernel arg modifications) require support from coreos-assembler writing
pointers in the System Area section to other places in the disc. With
the minimal ISO work, we'll need to able to find every file on the disc,
which would grow the complexity of these hacks further.

This patch adds a basic parser for the ISO9660 filesystem. It only
understands the minimal set of things we need, but it should provide a
good base for cleaning up our current approach, as well as for adding
minimal ISO support in a future patch.

As a way to exercise the code, add a new `coreos-installer iso inspect`
command which simply prints the volume descriptors and file contents of
any ISO.