iso: add `coreos-installer iso extract pxe`
This commands allows users to easily extract the PXE artifacts from
CoreOS live ISOs without having to mount them, which may require
additional privileges.

There are other CLI tools which can do this of course, but having it
integrated makes it much easier and it's trivial to support with the new
ISO9660 parser.

It also establishes the `extract` subcommand, to which we'll add a
`miniso` command for the minimal ISO work.