Makefile: Don't build at install time, and fix up RDCORE conditional
First, IMO as a general rule `make install` should *not* explicitly
depend on `make` i.e. `make all` because it's very common to use
`sudo make install` type flows which end up building the software
as root.  It's better to get an error if you just do `make install`
and require the user to invoke `make` first.

Secondly, let's not require the user to set the `RDCORE`
for both `make` and `make install`.  Instead, dynamically detect
if we built rdcore and conditionally install it.