io/verify: squash GPG stderr if we're not checking the result
Capture GPG's stderr and forward it to our stderr when checking GPG's
exit status.  If GpgReader is dropped without checking the exit status,
squash the stderr output.  This avoids implying that we've validated the
signature when we haven't, and also avoids spurious verification errors
if we fail a download or install for unrelated reasons.

Use eprint!() instead of io::stderr() to forward the output so that the
output will be automatically squashed in successful test cases.