source: add arch, platform, "Fedora CoreOS" to StreamLocation format
When starting an install from stream metadata, we were only reporting some
of the parameters used to select the source artifact.  Add the
architecture and platform.  In addition, if a custom stream metadata URL
is not specified, explicitly say "Fedora CoreOS" to avoid confusing users
who intend to install RHCOS.