download: report selected artifacts in download subcommand
The `download` subcommand was launching directly into the download, without
saying what it was downloading.  This could be confusing.  In particular,
it wasn't necessarily clear that we were downloading an FCOS image and not
an RHCOS image.  Print the location info as we do in `install`.

See https://bugzilla.redhat.com/show_bug.cgi?id=2009332.