install: inform user if filesystem labeled 'boot' is not unique
We shouldn't make this a hard error: if coreos-installer is being run not on
the target system (for example, writing to an external disk, or writing to a
guest VM disk from the host), it won't know what other disks will be
attached to the target system. In interactive installs, it'll be noticed and
acted upon in time. So this is just about opportunistically trying to save
users grief when possible (because it saves them another install iteration).

Signed-off-by: Nikita Dubrovskii <nikita@linux.ibm.com>