Add support for minimal ISO packing/unpacking
This command allows packing into the live ISO a minimal version of
itself which does not contain the root squashfs. The use case for
minimal ISOs has come up multiple times, notably in:

https://github.com/coreos/fedora-coreos-tracker/issues/661

This is a sort of intermediate approach where we don't officially ship a
new artifact, but allow users to derive the minimal ISO from the
official live ISO.

The way this works is similar to osmet. It requires a "packing" step at
build time which adds an osmet-like file (the miniso data file) on the
full ISO itself. The data file describes how to construct a minimal ISO
from the full ISO. It weighs in at a few kilobytes, so the impact of the
size of the live ISO is negligible. If we ever officially do ship a
minimal ISO, it also allows for the possibility of matching its checksum
exactly.

The method should be arch-independent and does not rely on e.g.
modifying ISO9660 structures or moving the GPT backup header, etc...

We support a `--rootfs-url` option which uses the `iso kargs` code to
inject `coreos.live.rootfs_url` since that's the primary use case. That
way, the minimal ISO can be prepared in a single step.

One of the primary goals is to ideally satisfy the needs of the Assisted
Installer (https://github.com/openshift/assisted-installer), though
there's more things needed before we get there (notably #545).