install: add no-op --config-file option; make device optional
The device might be specified in a config file or on the command line,
but any of those might not want to specify it.  Continue treating the
command-line option as required unless a config file is specified, but
also do a late check in case the config file doesn't provide one either.

While we're here, rename InstallConfig.device to dest_device so the
config file field name is more descriptive.