blockdev: add comments about using `lsblk --paths` for el7
This is a copy/paste from other instances of this. It's relevant as long
as we consider el7 a supported install platform (see
https://github.com/coreos/coreos-installer/issues/650). The particular
functions here aren't used yet at install time, but could in the future.

Anyway, we should probably instead have an `lsblk` command builder
helper where this lives just once.