scripts: print deprecation warning if /dev omitted from install_dev karg
We detect and work around the old syntax, and there are no plans to
stop supporting it.  However, the new syntax is more consistent, and while
our documentation tries to steer users toward it, examples of the old
syntax keep showing up.  Add a deprecation warning to encourage users to
update their configs.