live: add proxy struct between IsoConfig and Initrd
Add InitrdEmbedArea, similar to KargEmbedAreas, that manages the mapping
between Region and Initrd.  Because updating the Region contents requires
compressing, we want to defer it until we're ready to write.  The write
path doesn't take mutable self references, so encode the initrd into a
clone of the region and write from that instead.