live: add iso/pxe subcommands for embedding/wrapping NM keyfiles
An Ignition config for the live system might want to fetch remote
resources, and the system might need non-default network configuration
to do so.  This is possible in a limited way with kargs, but it's better
to expose the full flexibility of NetworkManager keyfiles.

Add iso subcommands to embed/extract/remove network settings, and pxe
subcommands to wrap/unwrap them.  These are analogous to the Ignition
embed mechanism, and in the case of ISO images they share the same initrd
and embed area.  NetworkManager keyfiles are deposited into
/etc/coreos-firstboot-network in the initrd, which is supported by the OS
as of https://github.com/coreos/fedora-coreos-config/pull/1371.

The show/unwrap subcommands are a bit different than the Ignition ones,
since there might be multiple keyfiles in an initrd.  By default, print
each keyfile to stdout, with a header listing the filename.  Also provide
a -C/--directory option to extract each keyfile to a directory.  Because
"show" doesn't completely capture the command's functionality, call the
ISO subcommand "iso network extract" instead.

Fixes https://github.com/coreos/coreos-installer/issues/545.