live: add initial "pxe customize" subcommand
Serialize and embed an empty Ignition config in a PXE initrd image.
Unlike the pxe wrap commands, this subcommand takes a CoreOS initramfs
and prepends it to the customizations.  This is consistent with the
"iso customize" semantics, and will allow us to validate that the OS
image supports the customization features we're going to use.

We verify that the input is a CoreOS live initrd, and that it hasn't
already been customized.  The former is done by checking for
/etc/coreos-live-initramfs, which was previously an internal API between
coreos-assembler and fedora-coreos-config; this makes that file's
existence into a stable API.

It's possible to implement in-place modification of the input, as
"iso customize" does by default, but implementing in-place reversion to
a pristine image would be tricky.  (We'd need to truncate the image to
the start of the concatenated CPIO archive that contains our
customizations.)  To avoid the footgun, omit in-place customization.
However, the subcommand syntax is parallel with "iso customize"
("-o -" is required for stdout) so we could add this later.