fixtures: add more live ISO fixtures
These were generated with a hacked-up cosa, but
https://github.com/coreos/coreos-assembler/pull/2619 should allow easier
fixture generation in the future.

Drop header modification code from iso-kargs.sh; we can just run the
script against multiple fixtures instead.  We don't currently have a
fixture without a coreKarg header, but that will happen soon, so add a
hack to test that case for now.