miniso: fix table validation failure on empty files
Empty files can be placed anywhere, even in the middle of another file,
which can cause table validation to fail.  But they also don't require us
to do anything, so just filter them out of the table.  For bookkeeping,
continue counting them as a success.

We don't currently have any empty files in the CoreOS ISO; discovered this
while making a test fixture.