cargo: disable LTO
LTO saves us 1.5 MB (11%) in the release binary with debug symbols
disabled.  On the other hand it's been an enormous pain.  The RHCOS
package disables LTO because of problems on s390x.  Quay and the Fedora
armv7hl builders apparently OOM building 0.12.0 with LTO.  And it's not a
default option in the first place.  Get rid of it.