cmdline: add man page generation
The idea is that `make docs` will generate the man pages via a `pack`
subcommand and we'll store them in the repo alongside the Markdown docs.
Since we don't need the generation code at runtime, gate it (and its deps)
behind an optional Cargo feature and have `make` enable the feature when
building with the debug profile.

These are mostly useful for distribution packages; the container doesn't
have a good way to expose them.  Also, the output isn't completely clean,
but at least we get man pages.