Makefile: strip `rdcore` DWARF in debug mode by default
When testing in debug mode, we don't want to copy the full ~180M
`rdcore` binary in the initramfs. In kola's default 1G of RAM, this can
cause `ENOSPC`. The debug mode is the default, and is most used by CI
and developers, i.e. those who'd use kola. In release mode, most used by
e.g. Koji/Brew, debug symbols gets split into the debuginfo package
anyway.

Tweak things so that by default in debug mode, we strip the binary. Do
it in a separate step instead of using `install --strip` because we only
want to strip DWARF info, but by default the full symbol table is
stripped and there's no way to customize it.

Allow the default to be overridden by explicitly passing `STRIP_RDCORE`.

This will have an effect on backtraces in case of crashes/panics.
Function names will still be printed, but filenames and line numbers
will be missing.