Warn on console= karg if --console could have been used instead
If one or more console= kargs are specified to the install or iso/pxe
customize subcommands, and Console can successfully parse all of those
kargs, warn the user that they might benefit from specifying them using
the console mechanism instead.

For legibility, wrap the message at 80 characters, but avoid wrapping in
the middle of the example argument lists.  Use the textwrap crate to do
this.  textwrap is already a clap dependency without default features, and
we also avoid non-default features here, so this doesn't increase the
binary weight.