blockdev: fail `install --save-partindex` on MBR-formatted disks
`install --save-partindex` and `install --save-partlabel` both search the GPT
for the specified partitions. While MBR doesn't have partition labels, it does
have partition indexes. Users might try to use `--save-partindex` with an MBR
disk, resulting in data loss when coreos-installer doesn't notice the
specified partitions.

To address this, we'll need to detect saved MBR partitions and convert them to
GPT partitions in the output. As a first step, if --save-partindex is specified
and the disk has an MBR but no GPT, fail install to avoid data loss.