workflows: cancel superseded Rust and RPM jobs
s390x tests are expensive, and when we're merging a lot of Dependabot PRs,
CI can get backed up running redundant tests on main.  We don't need to
continue running tests that have been superseded by new code, either in
main or in a PR branch, so add concurrency configuration to cancel them.