blockdev: use `-p` when calling `blkid`
By default, `blkid` will return cached data, which we don't want because
it might be stale. We need to use`-p` to make sure it directly probes
the block devices and bypasses the cache.

With `-p`, `blkid` requires passing the devices directly. Call it once
to gather the list of devices (we trust the cache enough for this) and
then again with `-p`.

We originally used `lsblk` here which uses the udev database, but this
was changed in https://github.com/coreos/coreos-installer/pull/813
because it failed to return the filesystem label in some instances. It
might be worth revisiting this at some point and find out if we were
just missing a `udevadm settle` somewhere.

See also: https://github.com/coreos/fedora-coreos-config/pull/2181#issuecomment-1397386896