tests/images: only test installed binary if env var set
In general, we want to test the binary built in the source tree and raise
an error if that binary is missing.  Historically we failed to notice
this case and fell back to testing the installed binary, if present.
CoreOS CI took advantage of this.

Fail if the coreos-installer binary isn't in the source tree, unless
COREOS_INSTALLER_TEST_INSTALLED_BINARY is set, in which case print a
warning and continue.

Followup to https://github.com/coreos/coreos-installer/pull/999#discussion_r1004688601.